backend/requirements/requirements.local.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cookiecutter==2.6.0
Django==5.1.3
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
Jinja2==3.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
urllib3==2.2.3
